Update README.md
Need to add screenshot of the interactive map (can’t do that right now as its in-progress)

How should we describe running the code? I described what the code does in the data section, but I wasn’t sure if we are using a makefile/what stage it is in

Do we want to add a future considerations section?